+++ src/Log/Logger.hs
@@ -3,7 +3,6 @@ module Log.Logger (
     Logger
   , mkLogger
   , mkBulkLogger
-  , withLogger
   , execLogger
   , waitForLogger
   , shutdownLogger
@@ -78,13 +77,6 @@ mkLogger name exec = mkLoggerImpl
 mkBulkLogger :: T.Text -> ([LogMessage] -> IO ()) -> IO () -> IO Logger
 mkBulkLogger = mkLoggerImpl
   newSQueueIO isEmptySQueue readSQueue writeSQueue (threadDelay 1000000)
-
--- | 'bracket' like execution of an 'IO' action, verifying all messages
--- are properly logged. See 'mkBulkLogger'.
-withLogger :: Logger -> (Logger -> IO r) -> IO r
-withLogger logger act = act logger `finally` cleanup
-  where
-    cleanup = waitForLogger logger >> shutdownLogger logger
 
 ----------------------------------------